runner/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-benchmark"
version = "0.1.0"
description = "Comprehensive LLM benchmark for software engineering, planning, product mind, and startup mind"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "litellm>=1.30",
    "pydantic>=2.0",
    "rich>=13.0",
    "tenacity>=8.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1",
    "mypy>=1.0",
]

[project.scripts]
llm-bench = "benchmark.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
benchmark = ["py.typed"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
